src/main/java/lab/squirrel/pojo/CallbackMsgText.java
package lab.squirrel.pojo;

import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlCData;
import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;

@JacksonXmlRootElement(localName = "xml")
public class CallbackMsgText extends CallbackMsg {
    @JacksonXmlCData
    @JacksonXmlProperty(localName = "Content")
    private String content;

    public CallbackMsgText() {
        setMsgType("text");
    }

    public String getContent() {
        return content;
    }

    public void setContent(String content) {
        this.content = content;
    }

}
